trtappUI/simple_app/requirements.txt
PySide6
pyserial
PyYAML
